Successfully moved Manvir's code onto uiScene
